// app/api/model/randomQuizDTO.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * 정션의 API 명세서
 * API 명세서
 * OpenAPI spec version: v1
 */
import type { RandomQuizDTOVideoType } from './randomQuizDTOVideoType';
import type { ChoiceResDTO } from './choiceResDTO';

/**
 * 퀴즈 랜덤 조회 응답 DTO
 */
export type RandomQuizDTO = {
  /** 영상 ID */
  id?: number;
  /** 퀴즈 문장 */
  sentence?: string;
  /** 정답 단어 */
  successWord?: string;
  /** 영상 URL */
  videoUrl?: string;
  /** 영상 타입 */
  videoType?: RandomQuizDTOVideoType;
  /** 영상과 관련된 대화/스크립트 */
  chat?: string;
  /** 비교 대상 단어 */
  compareWord?: string;
  /** 한국형 문장 즉 번역본 */
  korea?: string;
  /** 퀴즈 보기를 담은 리스트 */
  choices?: ChoiceResDTO[];
};